[dual-vm] commit vm2 for cli state vm2
# Conflicts:
#	Cargo.toml
