+++ cmd/starcoin/Cargo.toml
@@ -61,6 +61,7 @@ starcoin-vm2-storage = { workspace = true }
 starcoin-vm2-vm-types = { workspace = true }
 starcoin-vm2-abi-decoder = { workspace = true }
 starcoin-vm2-types = { workspace = true }
+starcoin-vm2-crypto = { workspace = true }
 
 [dev-dependencies]
 stest = { workspace = true }
